refactor(attention): split registry.ts into per-domain derive modules (gascity-dashboard-kc8e)
Moves the per-domain attention-derive logic out of registry.ts into
attention/derive/{activity,agents,beads,health,mail,maintainer,runs,shared}.ts,
dropping registry.ts from 1097 to 144 lines of wiring. Behavior-preserving: every
moved function body is byte-identical (same thresholds, severity, href, id
strings, freshness wrapping); the public export surface (createAttentionContributors,
GC_ESCALATION_LABEL, the *AttentionFacts types) is unchanged and all importers
compile; no test assertions changed. Closes gascity-dashboard-kc8e.
